requirements.txt
amqp==2.6.0
billiard==3.6.3.0
celery==4.4.6
click==7.1.2
etcd3==0.12.0
fastapi==0.59.0
future==0.18.2
grpcio==1.30.0
h11==0.9.0
httptools==0.1.1
kombu==4.6.11
librabbitmq==2.0.0
protobuf==3.12.2
pydantic==1.6
python-decouple==3.3
pytz==2020.1
six==1.15.0
starlette==0.13.4
tenacity==6.2.0
uvicorn==0.11.5
uvloop==0.14.0
vine==1.3.0
websockets==8.1
